+++ examples/json/server/main.go
@@ -29,7 +29,6 @@ import (
 
 	"go.uber.org/yarpc"
 	"go.uber.org/yarpc/encoding/json"
-	"go.uber.org/yarpc/transport"
 	"go.uber.org/yarpc/transport/http"
 	tch "go.uber.org/yarpc/transport/tchannel"
 
@@ -80,7 +79,7 @@ func main() {
 
 	dispatcher := yarpc.NewDispatcher(yarpc.Config{
 		Name: "keyvalue",
-		Inbounds: []transport.Inbound{
+		Inbounds: yarpc.Inbounds{
 			tch.NewInbound(channel, tch.ListenAddr(":28941")),
 			http.NewInbound(":24034"),
 		},